Improve `MaxHeightScrollArea` and add `NewNodeList`
The props passed to `MaxHeightScrollArea` were not passed to the
`ScrollArea` component correctly. They are now distributed between the
parent `div` and the `ScrollArea` component.

The `NewNodeList` component will be used to add new nodes to the canvas.
It will render as a list where nodes can be searched and selected.